combine cgo with rust ffi capabilities
