--- .dockerignore
@@ -1,3 +1,0 @@
-Dockerfile
-**.md
-.git
